+++ app/containers/AgentStats/index.js
@@ -15,40 +15,120 @@ import Radium from 'radium';
 
 import ErrorBoundary from 'components/ErrorBoundary';
 
+import Icon from 'components/Icon';
+
 import { deactivateToolbarStat } from 'containers/Toolbar/actions';
 import { selectToolbarStats } from 'containers/Toolbar/selectors';
 
 import Stat from 'components/Stat';
 import { selectAvailableStats } from './selectors';
 
+const styles = {
+  statsContainer: {
+    alignItems: 'center',
+    display: 'flex',
+    flexGrow: 1,
+    justifyContent: 'flex-end',
+    overflowX: 'hidden',
+    paddingTop: '300px',
+    marginTop: '-300px',
+  },
+  scrollStatsButton: {
+    padding: '15px 0',
+    zIndex: 2,
+    cursor: 'pointer',
+  },
+  scrollStatsLeft: {
+    marginRight: '-19px',
+    borderRadius: '0 5px 5px 0',
+  },
+  scrollStatsRight: {
+    marginLeft: '-19px',
+    borderRadius: '5px 0 0 5px',
+  },
+  ready: {
+    backgroundColor: '#072931',
+    ':hover': {
+      backgroundColor: '#093742',
+    },
+  },
+  notReady: {
+    backgroundColor: '#FE4565',
+    ':hover': {
+      backgroundColor: '#E43D5A',
+    },
+  },
+  statsScrollContainer: {
+    display: 'flex',
+    overflowX: 'hidden',
+    paddingTop: '300px',
+    marginTop: '-300px',
+  },
+};
+
 export class AgentStats extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      hoverIndex: -1,
+      statsHiddenLeft: false,
+      statsHiddenRight: false,
     };
   }
 
-  styles = {
-    statsContainer: {
-      alignItems: 'center',
-      display: 'flex',
-      minHeight: '54px',
-      flex: '1 1 auto',
-      justifyContent: 'flex-end',
-      order: '0',
-    },
-    button: {
-      cursor: 'pointer',
-    },
-  };
+  componentDidMount() {
+    window.addEventListener('resize', this.updateScrollItems);
+    this.updateScrollItems();
+  }
+
+  componentWillUnmount() {
+    window.removeEventListener('resize', this.updateScrollItems);
+  }
+
+  componentDidUpdate(prevProps) {
+    if (prevProps.toolbarStats.length !== this.props.toolbarStats.length) {
+      if (prevProps.toolbarStats.length < this.props.toolbarStats.length) {
+        // Scroll to the left to see newly added stat
+        this.statsScrollContainer.scrollLeft = 0;
+      }
+      this.updateScrollItems();
+    }
+  }
 
   removeStat = (stat) => {
     this.props.deactivateToolbarStat(stat);
-    this.setState({ hoverIndex: -1 });
   };
 
-  generateStat = (stat, index, array) => {
+  updateScrollItems = () => {
+    const scrollableWidth = this.statsScrollContainer.scrollWidth;
+    const viewableWidth = this.statsScrollContainer.clientWidth;
+    const scrollLeftPosition = this.statsScrollContainer.scrollLeft;
+
+    if (scrollableWidth >= viewableWidth) {
+      let statsHiddenLeft = false;
+      let statsHiddenRight = false;
+      if (scrollLeftPosition > 0) {
+        statsHiddenLeft = true;
+      }
+      if (scrollLeftPosition + viewableWidth < scrollableWidth) {
+        statsHiddenRight = true;
+      }
+      this.setState({ statsHiddenLeft, statsHiddenRight });
+    }
+  };
+
+  scrollLeft = () => {
+    this.statsScrollContainer.scrollLeft =
+      this.statsScrollContainer.scrollLeft - 104;
+    this.updateScrollItems();
+  };
+
+  scrollRight = () => {
+    this.statsScrollContainer.scrollLeft =
+      this.statsScrollContainer.scrollLeft + 104;
+    this.updateScrollItems();
+  };
+
+  generateStat = (stat, index) => {
     const userFriendlyName = this.props.availableStats[stat.statOption]
       .userFriendlyName;
     let key;
@@ -57,37 +137,110 @@ export class AgentStats extends React.Component {
     } else {
       key = stat.statOption + stat.statSource + stat.statAggregate;
     }
+    const leftOfStatIsVisible =
+      this.statsScrollContainer &&
+      index * 104 >= this.statsScrollContainer.scrollLeft;
+    const rightOfStatIsVisible =
+      this.statsScrollContainer &&
+      (index + 1) * 104 <=
+        this.statsScrollContainer.scrollLeft +
+          this.statsScrollContainer.clientWidth +
+          1;
+
     return (
-      <div
+      <Stat
         key={key}
-        onMouseOver={() => {
-          this.setState({ hoverIndex: index });
-        }}
-        onFocus={() => {
-          this.setState({ hoverIndex: index });
-        }}
-        onMouseLeave={() => {
-          this.setState({ hoverIndex: -1 });
-        }}
-      >
-        <Stat
-          key={key}
-          userFriendlyName={userFriendlyName}
-          hover={this.state.hoverIndex === index}
-          index={array.length - index}
-          stat={stat}
-          queues={this.props.queues}
-          removeStat={this.removeStat}
-          readyState={this.props.readyState}
-        />
-      </div>
+        userFriendlyName={userFriendlyName}
+        stat={stat}
+        index={index}
+        queues={this.props.queues}
+        removeStat={this.removeStat}
+        readyState={this.props.readyState}
+        canShowDetails={leftOfStatIsVisible && rightOfStatIsVisible}
+      />
     );
   };
 
   render() {
     return (
-      <div id="agent-stats" style={this.styles.statsContainer}>
-        {this.props.toolbarStats.map(this.generateStat)}
+      <div style={styles.statsContainer}>
+        {this.state.statsHiddenLeft &&
+          <div
+            style={{
+              paddingRight: '30px',
+              marginRight: '-30px',
+              background: `linear-gradient(to left, rgba(255,255,255,0) 0%, rgba(0,0,0,${this
+                .props.readyState === 'ready'
+                ? '0.7'
+                : '0.4'}) 100%)`,
+              zIndex: 1,
+              height: '100%',
+            }}
+          />}
+        {this.state.statsHiddenLeft &&
+          <div
+            id="scrollStatsLeft"
+            key="scrollStatsLeft"
+            style={[
+              styles.scrollStatsButton,
+              styles.scrollStatsLeft,
+              this.props.readyState === 'ready'
+                ? styles.ready
+                : styles.notReady,
+            ]}
+            onClick={this.scrollLeft}
+          >
+            <Icon
+              name="caret_white"
+              style={{
+                display: 'block',
+                transform: 'rotate(90deg)',
+              }}
+            />
+          </div>}
+        <div
+          id="agent-stats"
+          ref={(statsScrollContainer) => {
+            this.statsScrollContainer = statsScrollContainer;
+          }}
+          style={styles.statsScrollContainer}
+        >
+          {this.props.toolbarStats.map(this.generateStat)}
+        </div>
+        {this.state.statsHiddenRight &&
+          <div
+            style={{
+              paddingLeft: '30px',
+              marginLeft: '-30px',
+              background: `linear-gradient(to right, rgba(255,255,255,0) 0%, rgba(0,0,0,${this
+                .props.readyState === 'ready'
+                ? '0.7'
+                : '0.4'}) 100%)`,
+              zIndex: 1,
+              height: '100%',
+            }}
+          />}
+        {this.state.statsHiddenRight &&
+          <div
+            id="scrollStatsRight"
+            key="scrollStatsRight"
+            style={[
+              styles.scrollStatsButton,
+              styles.scrollStatsRight,
+              this.props.readyState === 'ready'
+                ? styles.ready
+                : styles.notReady,
+            ]}
+            onClick={this.scrollRight}
+          >
+            <Icon
+              name="caret_white"
+              style={{
+                display: 'block',
+                transform: 'rotate(-90deg)',
+              }}
+            />
+          </div>}
       </div>
     );
   }